Move Util source into NullAwayTestDataConstants and delete Util.java
Share UTIL_SOURCE from a single constant instead of duplicating it in
CoreTests, InitializationTests and ThriftTests, and add it inline to the
NullMarkednessTests and AutoSuggestTest cases that resolved Util from the
classpath, so the testdata resource can be removed.
